Remove voxel-clientmc for now since it is not released
Get it at https://github.com/deathcap/voxel-clientmc if you want
It worked with MC 1.7(?), kind of, but the player was kicked due to not
sending correct position updates. With 1.8 its dependency
node-minecraft-protocol is currently being updated. Not yet in a
publishable state for the NPM registry, and npm complains if a module is
not yet in NPM even if I give it a git URL, so removing it until then.
To help ease installation, GH-52

Also a kb-bindings-ui update to fix kb-bindings deprecation warning, deps